New feature - line renumbering
Renumbers all lines in the buffer, or only the lines in the active
region.  Gotos and gosubs in the whole visible buffer are updated if
their targets are renumbered.

Numbers non-empty lines even if they don't contain a number
initially.  This allows writing the whole program without numbers
initially, adding line numbers later.