Fix for renumbering regions
Renumbering can change the length of the buffer, rendering values
of (point) stored before the renumbering began obsolete.  This may
cause too many lines to be renumbered if we're only renumbering a
region.

So in case we are only renumbering a region, use a marker instead of a
plain point value.