Fix for changing buffer size in the non-region case.
Turns out renumbering a region was not the only case that got confused
by a changing buffer size.  Whole buffer renumbering had the same
issue.  So use a marker instead of a point value for point-end in all
cases.

This in turn means that we can skip the (eobp) test in the loop.