Fix indentation when decrease keyword is not at bol
Basic has : as a statement separator.  The keywords in
basic-decrease-indent-keywords-bol should be considered when they come
directly after a colon as well as when they occur at the beginning of
line.

Delay loops are commonly written on one line in C64 basic:
  for x = 1 to delay: next x

It was annoying to have that construct increase the indentation of
subsequenct lines.