Fix tabification bug when indenting code with line numbers (#28)
Previously, `basic-indent-line-to` removed the line number and called
`indent-line-to` which inserts tabs, presuming tab stops every
tab-width spaces. Then it prepended the line number to the beginning,
which had the side effect of changing how many spaces the first tab
was equivalent to.

For example, presuming a tab-width of 8, then (basic-indent-line-to 4)
uses four spaces, but (basic-indent-line-to 8) uses one tab. Since
spaces use relative positioning and tabs provide (relatively) absolute
positioning, prepending less than 8 characters of text would shift
over the four spaces but not change the tab.

This patch fixes that by simply untabifying the result from
`indent-line-to`; that is, all tabs are turned into the appropriate
number of spaces so everything is using relative positioning.